api/serializers/teacher_serializers.py
from rest_framework import serializers
from teacher.models.teacher import Teacher

class TeacherSerializer(serializers.ModelSerializer):
    class Meta:
        model = Teacher
        fields = "__all__"