requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.1
django-crispy-forms==2.3
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
PyMySQL==1.1.1
sqlparse==0.5.1
